backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0
openpyxl==3.1.2
reportlab==4.0.9
pyotp==2.9.0
qrcode==7.4.2
redis==5.0.1
django-redis==5.4.0
pytz==2024.1
sentry-sdk[django]==1.40.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
kombu==5.3.4
pytest-django==4.8.0
pytest-cov==4.1.0
drf-spectacular==0.27.2
